Added restcontroller

+++ src/main/java/ru/otus/spring/repository/BookRepository.java
@@ -11,7 +11,7 @@ public interface BookRepository extends JpaRepository<Book, Long> {
 
     Optional<Book> findByName(String name);
 
-    Long countById(Long bookId);
+    int countById(Long bookId);
 
     @EntityGraph(attributePaths = {"author", "genre"})
     //@Query("select b from Book b join fetch b.author join fetch b.genre")  //for inner join
